User query classes worked on
I started working on the query classes.
It will verify that passwords are the same for loggin in.

I also created a different class for pulling all of the user information when logging into a profile

Created a test class assosiated with the VerifyUser.kt . Currently my VerifyUser does not work as a method that returns a true or false.