Completion of two kotlin query methods
In this commit i finished writing the method codes for VerifyUser.kt & LoadUser.kt.

VerifyUser will return a Boolean when the passwords are the same or not.

LoadUsers currently only return the firstName, lastName, and organization of the user's profile. It returns it in a MutipleList<String>

I learned that when working with a kotlin "file" file type, the Class.forName will not work with the kotlin test classes.